part 4 finished

+++ controllers/blogs.js
@@ -1,5 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+/*const User = require('../models/user')
+const jwt = require('jsonwebtoken')*/
+const { userExtractor } = require('../utils/middleware')
 
 /* alkup.versio vertailun vuoksi
 blogsRouter.get('/', (request, response) => {
@@ -12,22 +15,49 @@ blogsRouter.get('/', (request, response) => {
 */
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  response.json(blog)
+})
+
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user //userExtractor
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    id: body.id,
+    user: user._id
+  })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  //console.log(request.user) userExtractor
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(400).json({ error: 'blog not found' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
+  // console.log(request.user) userExtractor
   const body = request.body
 
   const blog = {
